@@ -1,11 +1,11 @@
 {
-	"name": "@rize-io/active-win",
-	"version": "8.1.1-beta.24",
+	"name": "@rize-io/get-windows",
+	"version": "9.1.1",
 	"description": "Get metadata about the active window (title, id, bounds, owner, URL, etc)",
 	"license": "MIT",
 	"repository": {
 		"type": "git",
-		"url": "git+https://github.com/rize-io/active-win.git"
+		"url": "git+https://github.com/rize-io/get-windows.git"
 	},
 	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
@@ -13,20 +13,24 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
-	"engines": {
-		"node": ">=14"
+	"type": "module",
+	"exports": {
+		"types": "./index.d.ts",
+		"default": "./index.js"
 	},
-	"main": "./index",
-	"types": "./index.d.ts",
+	"main": "./index.js",
+	"sideEffects": false,
+	"engines": {
+		"node": ">=18.18"
+	},
 	"binary": {
-		"module_name": "node-active-win",
+		"module_name": "node-get-windows",
 		"module_path": "./lib/binding/napi-{napi_build_version}-{platform}-{libc}-{arch}",
-		"host": "https://github.com/rize-io/active-win/releases/download/",
+		"host": "https://github.com/rize-io/get-windows/releases/download/",
 		"remote_path": "v{version}",
 		"package_name": "napi-{napi_build_version}-{platform}-{libc}-{arch}.tar.gz",
 		"napi_versions": [
-			3,
-			6
+			9
 		]
 	},
 	"scripts": {
@@ -35,7 +39,7 @@
 		"build:windows:install": "node-pre-gyp install --fallback-to-build",
 		"build:windows": "node-pre-gyp build",
 		"build:windows:debug": "node-pre-gyp build --debug",
-		"build:macos": "swift build --configuration=release --arch arm64 --arch x86_64 && mv .build/apple/Products/Release/active-win main",
+		"build:macos": "swift build --configuration=release --arch arm64 --arch x86_64 && mv .build/apple/Products/Release/get-windows main",
 		"install": "node-pre-gyp install --fallback-to-build"
 	},
 	"files": [
@@ -53,10 +57,11 @@
 		"app",
 		"application",
 		"window",
-		"win",
+		"windows",
 		"active",
 		"focused",
 		"current",
+		"open",
 		"title",
 		"name",
 		"id",
@@ -76,25 +81,22 @@
 		"brave"
 	],
 	"devDependencies": {
-		"ava": "^2.4.0",
-		"tsd": "^0.14.0",
-		"xo": "^0.38.2",
+		"ava": "^6.1.2",
+		"tsd": "^0.31.0",
+		"xo": "^0.58.0",
 		"node-pre-gyp-github": "^1.4.4"
 	},
 	"optionalDependencies": {
-		"@mapbox/node-pre-gyp": "^1.0.10",
-		"node-addon-api": "^5.0.0",
-		"node-gyp": "^9.3.0"
+		"@mapbox/node-pre-gyp": "^1.0.11",
+		"node-addon-api": "^8.0.0",
+		"node-gyp": "^10.1.0"
 	},
 	"peerDependencies": {
-		"node-gyp": "^9.3.0"
+		"node-gyp": "^10.1.0"
 	},
 	"peerDependenciesMeta": {
 		"node-gyp": {
 			"optional": true
 		}
-	},
-	"ava": {
-		"verbose": true
 	}
 }
