@@ -1,17 +1,17 @@
-'use strict';
-const {promisify} = require('util');
-const fs = require('fs');
-const childProcess = require('child_process');
+import process from 'node:process';
+import {promisify} from 'node:util';
+import fs from 'node:fs';
+import childProcess from 'node:child_process';
 
 const execFile = promisify(childProcess.execFile);
 const readFile = promisify(fs.readFile);
 const readlink = promisify(fs.readlink);
 
-const xpropBin = 'xprop';
-const xwininfoBin = 'xwininfo';
-const xpropActiveArgs = ['-root', '\t$0', '_NET_ACTIVE_WINDOW'];
-const xpropOpenArgs = ['-root', '_NET_CLIENT_LIST_STACKING'];
-const xpropDetailsArgs = ['-id'];
+const xpropBinary = 'xprop';
+const xwininfoBinary = 'xwininfo';
+const xpropActiveArguments = ['-root', '\t$0', '_NET_ACTIVE_WINDOW'];
+const xpropOpenArguments = ['-root', '_NET_CLIENT_LIST_STACKING'];
+const xpropDetailsArguments = ['-id'];
 
 const processOutput = output => {
 	const result = {};
@@ -35,8 +35,8 @@ const parseLinux = ({stdout, boundsStdout, activeWindowId}) => {
 
 	const windowIdProperty = 'WM_CLIENT_LEADER(WINDOW)';
 	const resultKeys = Object.keys(result);
-	const windowId = (resultKeys.indexOf(windowIdProperty) > 0 &&
-		Number.parseInt(result[windowIdProperty].split('#').pop(), 16)) || activeWindowId;
+	const windowId = (resultKeys.indexOf(windowIdProperty) > 0
+		&& Number.parseInt(result[windowIdProperty].split('#').pop(), 16)) || activeWindowId;
 
 	const processId = Number.parseInt(result['_NET_WM_PID(CARDINAL)'], 10);
 
@@ -50,14 +50,14 @@ const parseLinux = ({stdout, boundsStdout, activeWindowId}) => {
 		id: windowId,
 		owner: {
 			name: JSON.parse(result['WM_CLASS(STRING)'].split(',').pop()),
-			processId
+			processId,
 		},
 		bounds: {
 			x: Number.parseInt(bounds['Absolute upper-left X'], 10),
 			y: Number.parseInt(bounds['Absolute upper-left Y'], 10),
 			width: Number.parseInt(bounds.Width, 10),
-			height: Number.parseInt(bounds.Height, 10)
-		}
+			height: Number.parseInt(bounds.Height, 10),
+		},
 	};
 };
 
@@ -69,13 +69,11 @@ const getMemoryUsageByPid = async pid => {
 };
 
 const getMemoryUsageByPidSync = pid => {
-	const statm = require('fs').readFileSync(`/proc/${pid}/statm`, 'utf8');
+	const statm = fs.readFileSync(`/proc/${pid}/statm`, 'utf8');
 	return Number.parseInt(statm.split(' ')[1], 10) * 4096;
 };
 
-const getPathByPid = pid => {
-	return readlink(`/proc/${pid}/exe`);
-};
+const getPathByPid = pid => readlink(`/proc/${pid}/exe`);
 
 const getPathByPidSync = pid => {
 	try {
@@ -85,18 +83,18 @@ const getPathByPidSync = pid => {
 
 async function getWindowInformation(windowId) {
 	const [{stdout}, {stdout: boundsStdout}] = await Promise.all([
-		execFile(xpropBin, [...xpropDetailsArgs, windowId], {env: {...process.env, LC_ALL: 'C.utf8'}}),
-		execFile(xwininfoBin, [...xpropDetailsArgs, windowId])
+		execFile(xpropBinary, [...xpropDetailsArguments, windowId], {env: {...process.env, LC_ALL: 'C.utf8'}}),
+		execFile(xwininfoBinary, [...xpropDetailsArguments, windowId]),
 	]);
 
 	const data = parseLinux({
-		windowId,
+		activeWindowId: windowId,
 		boundsStdout,
-		stdout
+		stdout,
 	});
 	const [memoryUsage, path] = await Promise.all([
 		getMemoryUsageByPid(data.owner.processId),
-		getPathByPid(data.owner.processId).catch(() => {})
+		getPathByPid(data.owner.processId).catch(() => {}),
 	]);
 	data.memoryUsage = memoryUsage;
 	data.owner.path = path;
@@ -104,22 +102,22 @@ async function getWindowInformation(windowId) {
 }
 
 function getWindowInformationSync(windowId) {
-	const stdout = childProcess.execFileSync(xpropBin, [...xpropDetailsArgs, windowId], {encoding: 'utf8', env: {...process.env, LC_ALL: 'C.utf8'}});
-	const boundsStdout = childProcess.execFileSync(xwininfoBin, [...xpropDetailsArgs, windowId], {encoding: 'utf8'});
+	const stdout = childProcess.execFileSync(xpropBinary, [...xpropDetailsArguments, windowId], {encoding: 'utf8', env: {...process.env, LC_ALL: 'C.utf8'}});
+	const boundsStdout = childProcess.execFileSync(xwininfoBinary, [...xpropDetailsArguments, windowId], {encoding: 'utf8'});
 
 	const data = parseLinux({
-		windowId,
+		activeWindowId: windowId,
 		boundsStdout,
-		stdout
+		stdout,
 	});
 	data.memoryUsage = getMemoryUsageByPidSync(data.owner.processId);
 	data.owner.path = getPathByPidSync(data.owner.processId);
 	return data;
 }
 
-module.exports = async () => {
+export async function activeWindow() {
 	try {
-		const {stdout: activeWindowIdStdout} = await execFile(xpropBin, xpropActiveArgs);
+		const {stdout: activeWindowIdStdout} = await execFile(xpropBinary, xpropActiveArguments);
 		const activeWindowId = getActiveWindowId(activeWindowIdStdout);
 
 		if (!activeWindowId) {
@@ -130,11 +128,11 @@ module.exports = async () => {
 	} catch {
 		return undefined;
 	}
-};
+}
 
-module.exports.sync = () => {
+export function activeWindowSync() {
 	try {
-		const activeWindowIdStdout = childProcess.execFileSync(xpropBin, xpropActiveArgs, {encoding: 'utf8'});
+		const activeWindowIdStdout = childProcess.execFileSync(xpropBinary, xpropActiveArguments, {encoding: 'utf8'});
 		const activeWindowId = getActiveWindowId(activeWindowIdStdout);
 
 		if (!activeWindowId) {
@@ -145,11 +143,12 @@ module.exports.sync = () => {
 	} catch {
 		return undefined;
 	}
-};
+}
 
-module.exports.getOpenWindows = async () => {
+export async function openWindows() {
 	try {
-		const {stdout: openWindowIdStdout} = await execFile(xpropBin, xpropOpenArgs);
+		const {stdout: openWindowIdStdout} = await execFile(xpropBinary, xpropOpenArguments);
+
 		// Get open windows Ids
 		const windowsIds = openWindowIdStdout.split('#')[1].trim().replace('\n', '').split(',');
 
@@ -167,11 +166,11 @@ module.exports.getOpenWindows = async () => {
 	} catch {
 		return undefined;
 	}
-};
+}
 
-module.exports.getOpenWindowsSync = () => {
+export function openWindowsSync() {
 	try {
-		const openWindowIdStdout = childProcess.execFileSync(xpropBin, xpropOpenArgs, {encoding: 'utf8'});
+		const openWindowIdStdout = childProcess.execFileSync(xpropBinary, xpropOpenArguments, {encoding: 'utf8'});
 		const windowsIds = openWindowIdStdout.split('#')[1].trim().replace('\n', '').split(',');
 
 		if (!windowsIds || windowsIds.length === 0) {
@@ -190,4 +189,4 @@ module.exports.getOpenWindowsSync = () => {
 		console.log(error);
 		return undefined;
 	}
-};
+}
