@@ -1,10 +1,12 @@
-'use strict';
-const path = require('path');
-const {promisify} = require('util');
-const childProcess = require('child_process');
+import path from 'node:path';
+import {promisify} from 'node:util';
+import childProcess from 'node:child_process';
+import {fileURLToPath} from 'node:url';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const execFile = promisify(childProcess.execFile);
-const bin = path.join(__dirname, '../main');
+const binary = path.join(__dirname, '../main');
 
 const parseMac = stdout => {
 	try {
@@ -24,30 +26,34 @@ const getArguments = options => {
 		return [];
 	}
 
-	const args = [];
-	if (options.screenRecordingPermission === false) {
-		args.push('--no-screen-recording-permission');
+	const arguments_ = [];
+	if (options.accessibilityPermission === false) {
+		arguments_.push('--no-accessibility-permission');
 	}
 
-	return args;
+	if (options.screenRecordingPermission === false) {
+		arguments_.push('--no-screen-recording-permission');
+	}
+
+	return arguments_;
 };
 
-module.exports = async options => {
-	const {stdout} = await execFile(bin, getArguments(options));
+export async function activeWindow(options) {
+	const {stdout} = await execFile(binary, getArguments(options));
 	return parseMac(stdout);
-};
+}
 
-module.exports.sync = options => {
-	const stdout = childProcess.execFileSync(bin, getArguments(options), {encoding: 'utf8'});
+export function activeWindowSync(options) {
+	const stdout = childProcess.execFileSync(binary, getArguments(options), {encoding: 'utf8'});
 	return parseMac(stdout);
-};
+}
 
-module.exports.getOpenWindows = async options => {
-	const {stdout} = await execFile(bin, [...getArguments(options), '--open-windows-list']);
+export async function openWindows(options) {
+	const {stdout} = await execFile(binary, [...getArguments(options), '--open-windows-list']);
 	return parseMac(stdout);
-};
+}
 
-module.exports.getOpenWindowsSync = options => {
-	const stdout = childProcess.execFileSync(bin, [...getArguments(options), '--open-windows-list'], {encoding: 'utf8'});
+export function openWindowsSync(options) {
+	const stdout = childProcess.execFileSync(binary, [...getArguments(options), '--open-windows-list'], {encoding: 'utf8'});
 	return parseMac(stdout);
-};
+}
