@@ -1,11 +1,34 @@
-const path = require('path');
-const bindingPath = path.join(__dirname, 'windows-binding.js');
-const addon = require(bindingPath);
+import path from 'node:path';
+import fs from 'node:fs';
+import {fileURLToPath} from 'node:url';
+import {createRequire} from 'node:module';
+import preGyp from '@mapbox/node-pre-gyp';
 
-module.exports = async () => addon.getActiveWindow();
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
-module.exports.getOpenWindows = async () => addon.getOpenWindows();
+const getAddon = () => {
+	const require = createRequire(import.meta.url);
 
-module.exports.sync = addon.getActiveWindow;
+	const bindingPath = preGyp.find(path.resolve(path.join(__dirname, '../package.json')));
 
-module.exports.getOpenWindowsSync = addon.getOpenWindows;
+	return (fs.existsSync(bindingPath)) ? require(bindingPath) : {
+		getActiveWindow() {},
+		getOpenWindows() {},
+	};
+};
+
+export async function activeWindow() {
+	return getAddon().getActiveWindow();
+}
+
+export function activeWindowSync() {
+	return getAddon().getActiveWindow();
+}
+
+export function openWindows() {
+	return getAddon().getOpenWindows();
+}
+
+export function openWindowsSync() {
+	return getAddon().getOpenWindows();
+}
