@@ -117,8 +117,9 @@ func getWindowInformation(window: [String: Any], windowOwnerPID: pid_t) -> [Stri
 		"memoryUsage": window[kCGWindowMemoryUsage as String] as? Int ?? 0
 	]
 
-	// Only run the AppleScript if active window is a compatible browser.
+	// Run the AppleScript to get the URL if the active window is a compatible browser and accessibility permissions are enabled.
 	if
+		!disableAccessibilityPermission,
 		let bundleIdentifier = app.bundleIdentifier,
 		let script = getActiveBrowserTabURLAppleScriptCommand(bundleIdentifier),
 		let windowData = runAppleScript(source: script)
@@ -132,18 +133,24 @@ func getWindowInformation(window: [String: Any], windowOwnerPID: pid_t) -> [Stri
 	return output
 }
 
+let disableAccessibilityPermission = CommandLine.arguments.contains("--no-accessibility-permission")
 let disableScreenRecordingPermission = CommandLine.arguments.contains("--no-screen-recording-permission")
 let enableOpenWindowsList = CommandLine.arguments.contains("--open-windows-list")
 
-// Show accessibility permission prompt if needed. Required to get the complete window title.
-if !AXIsProcessTrustedWithOptions(["AXTrustedCheckOptionPrompt": true] as CFDictionary) {
-	print("active-win requires the accessibility permission in “System Settings › Privacy & Security › Accessibility”.")
-	exit(1)
+// Show accessibility permission prompt if needed. Required to get the URL of the active tab in browsers.
+if !disableAccessibilityPermission {
+	if !AXIsProcessTrustedWithOptions(["AXTrustedCheckOptionPrompt": true] as CFDictionary) {
+		print("get-windows requires the accessibility permission in “System Settings › Privacy & Security › Accessibility”.")
+		exit(1)
+	}
 }
 
 // Show screen recording permission prompt if needed. Required to get the complete window title.
-if !disableScreenRecordingPermission && !hasScreenRecordingPermission() {
-	print("active-win requires the screen recording permission in “System Settings › Privacy & Security › Screen Recording”.")
+if
+	!disableScreenRecordingPermission,
+	!hasScreenRecordingPermission()
+{
+	print("get-windows requires the screen recording permission in “System Settings › Privacy & Security › Screen Recording”.")
 	exit(1)
 }
 
